Warn for join before A6 and split before knight
While A6 can start with two legs, it is not common. So if there is a
join to two legs declared (A1d), print a warning to check if the coach
meant a single leg lift instead (A1a, A1c).

Likewise, split (F1a) before is knight is possible, but F1b (going to
knight) is way more likely since it is a "gimmie" since you have to go
to a knight position to do any of the actions that start from a knight
position.

Closes #9, closes #10